introduce safe ts mgr
