+++ dbms/src/Storages/KVStore/tests/gtest_learner_read.cpp
@@ -125,8 +125,8 @@ try
     // region_202 has no safe ts info, can not stale read
     // region_203 read index cache exist in the `mvcc_query_info`
     RegionTable region_table(global_ctx);
-    region_table.updateSafeTS(region_id_200, /*leader_safe_ts*/ 10005, /*self_safe_ts*/ 10005);
-    region_table.updateSafeTS(region_id_201, /*leader_safe_ts*/ 9995, /*self_safe_ts*/ 9995);
+    region_table.safeTsMgr().updateSafeTS(region_id_200, /*leader_safe_ts*/ 10005, /*self_safe_ts*/ 10005);
+    region_table.safeTsMgr().updateSafeTS(region_id_201, /*leader_safe_ts*/ 9995, /*self_safe_ts*/ 9995);
 
     RegionsReadIndexResult read_index_result;
     auto requests = buildBatchReadIndex(worker, region_table, snapshot, read_index_result);
